backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic[email]==2.10.3
pydantic-settings==2.6.1
python-dateutil==2.9.0
pytz==2024.2
sqlalchemy==2.0.36
aiofiles==24.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.17
slowapi==0.1.9
